backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "In Plain Sight gallery backend"
authors = ["Ariel Pillemer <pillemer@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.125.0"
uvicorn = "^0.38.0"
strawberry-graphql = "^0.287.3"
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
ruff = "^0.14.0"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
